+++ src/app/app.component.ts
@@ -1,10 +1,12 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { CenterBody } from './center-body';
+import { CenterBody } from './center-body/center-body';
+import { ComponentHeaderComponent } from "./component-header/component-header.component";
+
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [RouterOutlet, CenterBody],
+  imports: [RouterOutlet, CenterBody, ComponentHeaderComponent],
   templateUrl: './app.component.html',
   styleUrl: './app.component.css'
 })